Move blog components to route-specific directory
Renamed ContentTable, HeadingWithImage, Spacer, Subheading, and TextWithImage components to src/app/[lang]/blog/[slug]/_components for better organization. Updated imports in affected pages to reflect new paths.
