+++ src/app/[lang]/blog/[slug]/_components/ContentTable.tsx
