+++ src/app/[lang]/blog/[slug]/_components/HeadingWithImage.tsx
