+++ src/app/[lang]/blog/[slug]/_components/TextWithImage.tsx
